Move FormXmlLoader to new metadata namespace
