+++ src/Sulu/Bundle/AdminBundle/Tests/Unit/Metadata/FormMetadata/data/form_invalid_root_tag.xml
